requirements.txt
Flask==3.0.0
selenium==4.18.1
python-dotenv==1.0.1
google-cloud-secret-manager==2.18.1